move extension to new file
